feat: extract parsing and dataset modules
